Add comments and move old code to it's own module
